+++ swift-sdk/Core/Constants.swift
@@ -96,6 +96,7 @@ enum Const {
         static let visitorConsentTimestamp = "itbl_visitor_consent_timestamp"
         static let isNotificationsEnabled = "itbl_isNotificationsEnabled"
         static let hasStoredNotificationSetting = "itbl_hasStoredNotificationSetting"
+        static let keychainMigrationCompleted = "itbl_keychain_migration_completed"
 
         static let attributionInfoExpiration = 24
     }